+++ redis/deprecated/redisbp/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
         "hooks.go",
         "monitored_client.go",
     ],
-    importpath = "github.com/reddit/baseplate.go/redisbp",
+    importpath = "github.com/reddit/baseplate.go/redis/deprecated/redisbp",
     visibility = ["//visibility:public"],
     deps = [
         "//errorsbp:go_default_library",
@@ -38,7 +38,7 @@ go_test(
         "//mqsend:go_default_library",
         "//thriftbp:go_default_library",
         "//tracing:go_default_library",
-        "@com_github_alicebob_miniredis//:go_default_library",
+        "@com_github_alicebob_miniredis_v2//:go_default_library",
         "@com_github_go_redis_redis_v7//:go_default_library",
         "@com_github_opentracing_opentracing_go//:go_default_library",
         "@in_gopkg_yaml_v2//:go_default_library",
