+++ redis/deprecated/redisbp/doc.go
@@ -2,4 +2,6 @@
 //
 // See https://pkg.go.dev/github.com/go-redis/redis/v7 for documentation for
 // go-redis
+// redisbp is deprecated and will be removed in a future version, use redisx
+//  and redispipebp instead.
 package redisbp
