+++ redis/deprecated/redisbp/example_monitored_client_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/go-redis/redis/v7"
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/reddit/baseplate.go/metricsbp"
-	"github.com/reddit/baseplate.go/redisbp"
+	"github.com/reddit/baseplate.go/redis/deprecated/redisbp"
 	"github.com/reddit/baseplate.go/tracing"
 )
