+++ redis/deprecated/redisbp/monitored_client_test.go
@@ -6,10 +6,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/alicebob/miniredis"
+	"github.com/alicebob/miniredis/v2"
 	"github.com/go-redis/redis/v7"
 	"github.com/reddit/baseplate.go/mqsend"
-	"github.com/reddit/baseplate.go/redisbp"
+	"github.com/reddit/baseplate.go/redis/deprecated/redisbp"
 	"github.com/reddit/baseplate.go/tracing"
 )
